requirements.txt
click
openai
chromadb
python-dotenv
langchain
sentence-transformers
RAGchain~=0.2.0
gradio
